fixed select for larger canvas

+++ canvas.cpp
@@ -1040,7 +1040,7 @@ void cCanvas::Editor() {
 
                 for (int tileY = startTileY; tileY < endTileY; tileY++) {
                     for (int tileX = startTileX; tileX < endTileX; tileX++)
-                        selectedIndexes.insert((uint16_t)(tileX + tileY * width));
+                        selectedIndexes.insert((int64_t)(tileX + tileY * width));
                 }
 
                 if (selectedIndexes.size() <= 1) selectedIndexes.clear();
